Move RobotProperty and RobotPropertyKind to rmf_site_editor
